requirements.txt
python-dotenv>=1.0.0
langchain>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
langchain-text-splitters>=0.3.0
langchain-core>=0.3.0
langchain-salesforce>=0.1.1
streamlit